Added new enemyLevel class to inherit from level
